requirements.txt
yfinance
pandas
psycopg2-binary
requests